Fragment experiment
Add a special case in `createElement` for a JSX tag named `<frag>`. When
`createElement` is called on a `<frag>`, it will create a
`ReactFragment` instead of a `ReactElement`. For example, the JSX:

```
<div>
  <h2>greetings!</h2>
  <frag>
    <p>Foo</p>
    <p>Bar</p>
  </frag>
</div>
```

Would result in the HTML:

```
<div data-reactid=".0">
  <h2 data-reactid=".0.0">
    greetings!
  </h2>
  <p data-reactid=".0.1:$frag-0">
    Foo
  </p>
  <p data-reactid=".0.1:$frag-1">
    Bar
  </p>
</div>
```

This allows fragments to be seamlessly created using a tag in JSX. While
fragments currently cannot be returned directly from components, they
can still be used in other JSX expressions.