Fragment experiment
Add a special case in `createElement` for a JSX tag named `<frag>`. When
`createElement` is called on a `<frag>`, it will simply return the
children of the `<frag/>` as an array. Thus, the following JSX:

```
<div>
  <h2>greetings!</h2>
  <frag>
    <p>Foo</p>
    <p>Bar</p>
  </frag>
</div>
```

Would result in the DOM:

```
<div data-reactid=".0">
  <h2 data-reactid=".0.0">
    greetings!
  </h2>
  <p data-reactid=".0.1:0">
    Foo
  </p>
  <p data-reactid=".0.1:1">
    Bar
  </p>
</div>
```

While these fragments currently cannot be returned directly from components,
they can still be used in other JSX expressions.